2nd pass of style integration and code cleanup
