+++ src/components/Presenting.jsx
@@ -1,17 +1,15 @@
-import React from 'react'
+import {Viewer, Worker} from '@react-pdf-viewer/core'
 import './PDF_viewer.css'
-import { Viewer, Worker } from '@react-pdf-viewer/core'
-import { defaultLayoutPlugin } from '@react-pdf-viewer/default-layout'
 import '@react-pdf-viewer/core/lib/styles/index.css'
 import '@react-pdf-viewer/default-layout/lib/styles/index.css'
+import {defaultLayoutPlugin} from '@react-pdf-viewer/default-layout'
 
-export default function ViewPDF(params) {
+export default function Presenting(params) {
 
-    const newplugin = defaultLayoutPlugin()
+    const newplugin= defaultLayoutPlugin()
 
     return (
-        <div>
-            <div className='pdf-container'>
+        <div className='pdf-container'>
                 <Worker workerUrl="https://unpkg.com/pdfjs-dist@2.15.349/build/pdf.worker.min.js">
                     {params.pdfFile && <>
                         <Viewer fileUrl={params.pdfFile} plugins={[newplugin]} />
@@ -19,7 +17,5 @@ export default function ViewPDF(params) {
                     {!params.pdfFile && <>No PDF</>}
                 </Worker>
             </div>
-        </div>
-
-    )
-};
+    );
+    }
