+++ src/components/styles.css
@@ -3,7 +3,6 @@
   display: inline-block;
   font-weight: 400;
   text-align: center;
-  /* white-space: nowrap; */
   vertical-align: middle;
   user-select: none;
   border: 1px solid transparent;
@@ -20,7 +19,6 @@
 .custom-button:hover {
   color: #0056b3;
   text-decoration: none;
-  /* background-color: #e6e6e6; */
   cursor: pointer;
 }
 
@@ -31,15 +29,11 @@
 
 .custom-button:active {
   color: #0062cc;
-  /* background-color: #d6d6d6; */
-  /* border-color: #bdbdbd; */
 }
 
 .custom-button:disabled {
   color: #6c757d;
   pointer-events: none;
-  /* background-color: #e9ecef; */
-  /* border-color: #dee2e6; */
 }
 
 /* General styles for containers */
@@ -138,99 +132,91 @@
 
 /* Dark theme styles */
 .custom-navbar.dark-theme {
-  /* background-color: rgb(32, 33, 35); */
-  /* color: rgb(178, 178, 178); */
   box-shadow: 1px 1px 5px rgb(0, 0, 0);
 }
 
 .custom-navbar.dark-theme .custom-button {
-  /* color: rgb(255, 255, 255); */
-  /* background-color: rgb(32, 33, 35); */
-  /* border-color: rgb(178, 178, 178); */
   box-shadow: 1px 1px 5px rgb(0, 0, 0);
 }
 
 /* Light theme styles */
 .custom-navbar.light-theme {
-  /* background-color: white; */
-  /* color: rgb(32, 33, 35); */
   box-shadow: 1px 1px 5px rgba(0, 0, 0, 0.3);
 }
 
 .custom-navbar.light-theme .custom-button {
-  /* color: rgb(32, 33, 35); */
-  /* background-color: white; */
-  /* border-color: white; */
   box-shadow: 1px 1px 5px rgba(0, 0, 0, 0.3);
 }
 
 .custom-button.dark-theme {
-  /* color: rgb(255, 255, 255); */
-  /* background-color: rgb(32, 33, 35); */
-  /* border-color: rgb(178, 178, 178); */
   box-shadow: 1px 1px 5px rgb(0, 0, 0);
 }
 
 .custom-button.light-theme {
-  /* color: rgb(32, 33, 35); */
-  /* background-color: white; */
-  /* border-color: white; */
   box-shadow: 1px 1px 5px rgba(0, 0, 0, 0.3);
 }
 
 
 .custom-button:hover.light-theme {
-  /* color: rgb(0, 0, 0); */
   text-decoration: none;
-  /* background-color: #e6e6e6; */
-  /* border-color: #ddd; */
   cursor: pointer;
   box-shadow: 0px 0px 5px rgba(0, 0, 0, 0.645);
 }
 
 .custom-button:hover.dark-theme {
-  /* color: rgb(255, 255, 255); */
   text-decoration: none;
-  /* background-color: rgb(40, 41, 44); */
-  /* border-color: #ddd; */
   cursor: pointer;
   box-shadow: 0px 0px 5px rgba(255, 255, 255, 0.61);
 }
 
 
 .hcontainer.dark-theme {
-  /* background-color: rgb(32, 33, 35); */
-  /* color: rgb(255, 255, 255); */
   box-shadow: 1px 1px 5px rgb(0, 0, 0);
 }
 
 .hcontainer.light-theme {
-  /* background-color: white; */
-  /* color: rgb(32, 33, 35); */
   box-shadow: 1px 1px 5px rgba(0, 0, 0, 0.3);
 }
 
 .vcontainer.dark-theme {
-  /* background-color: rgb(32, 33, 35); */
-  /* color: rgb(255, 255, 255); */
   box-shadow: 1px 1px 5px rgb(0, 0, 0);
 }
 
 .vcontainer.light-theme {
-  /* background-color: white; */
-  /* color: rgb(32, 33, 35); */
   box-shadow: 1px 1px 5px rgba(0, 0, 0, 0.3);
 }
 
-.home-container.dark-theme {
-  /* background-color: rgb(32, 33, 35); */
-  /* color: rgb(255, 255, 255); */
-  /* box-shadow: 1px 1px 5px rgb(0, 0, 0); */
+/* General styles for custom-text-input */
+.custom-text-input {
+  display: inline-block;
+  font-weight: 400;
+  text-align: center;
+  vertical-align: middle;
+  user-select: none;
+  border: 1px solid transparent;
+  padding: 0.375rem 0.75rem;
+  margin: 1px;
+  font-size: 3rem;
+  line-height: 1.5;
+  border-radius: 0.25rem;
+  box-shadow: 1px 1px 4px rgba(0, 0, 0, 0.5);
+  background-color: inherit;
+  color: inherit;
 }
 
-.home-container.light-theme {
-  /* background-color: white; */
-  /* color: rgb(32, 33, 35); */
-  /* box-shadow: 1px 1px 5px rgba(0, 0, 0, 0.3); */
+.custom-text-input:hover {
+  background-color: #f5f5f5;
+  cursor: text;
 }
 
+.custom-text-input:focus {
+  outline: 0;
+  box-shadow: 0 0 0 0.2rem rgba(0, 123, 255, 0.25);
+  background-color: #f5f5f5;
+}
+
+.custom-text-input:disabled {
+  background-color: #e9ecef;
+  color: #6c757d;
+  pointer-events: none;
+}
